fix: Monitoring Dashboard Workloads Controller Detail 컴포넌트로 합침 & MiniUsageChart 수정
